chore: set auth header for EV and ER plugins
