+++ .env.development
@@ -1,1 +1,2 @@
 DHIS2_BASE_URL="http://localhost:8080"
+DHIS2_AUTHORIZATION="Base64 encoded password"
